feat: returns operations content
